refactor: extract drop events from AgentWorkspace
