+++ src/components/agent/use-agent-drop-events-types.ts
@@ -1,0 +1,11 @@
+import type { AgentAttachment, ImageSafeModeConsentEventPayload } from "./agent-workspace-models";
+
+export type UseAgentDropEventsDependencies = {
+  handleAgentImageSafeModeConsentEvent: (
+    payload?: ImageSafeModeConsentEventPayload,
+  ) => Promise<void>;
+  importDroppedFilePaths: (
+    paths: string[],
+    options?: { onImported?: (attachments: AgentAttachment[]) => void },
+  ) => Promise<boolean>;
+};
